MINTEMP checking is no longer dependent on ambient temperature
Instead of constantly checking bed/hotend temp vs. MINTEMP while ambient>MINTEMP,
now we only take the "ambient<MINTEMP" path where we only check bed/hotend temps
vs. MINTEMP *while the heaters are engaged*.

So we can boot up fine in a cold room and sit forever without getting a MINTEMP
error. But if the target temp is > MINTEMP and the heater has been on for at least
15 seconds (50 seconds for the bed), we'll start enforcing MINTEMP and will die
with a MINTEMP error if temps drop below.